Clean up the handling of "ver" among messages.
It is not a requirement of keep them in sync among the code base.
In fact, it is better to see them as a minimum version and to increase
only messages that are affected. Increases the ver will split the
population. That can be intentional to get rid of old clients - like
in this commit - but it should be done only if needed.

Also, changed the type from string to numbers to simplify the processing
on the server in the future.