Request: fix firefox page store (#123)
On Firefox, in case a main_frame got redirected, the webRequest events were updating the page url to the intermediate value. That was breaking the previousPage logic and was causing stats to be attributed to the wrong page object.

In general updating urls based on webrequest is tricky as the request may fail and this is what happens during redirects. So alternative option would be to do it on onCompleted events only, but on the other hand, webNavigation and tabs.onUpdated do the book keeping just fine.