Retry mechanism for recoverable jobs
Jobs are placed in a "retryable" queue. If a job of the same type
succeeds, one job from that queue will be put into the "ready" queue.

By default, jobs will be dropped after they failed three times. Jobs that
failed with an unclassified or permanent error will never be retried.