Mitigate midnight spike of anti-tracking signals
Since sending via the jobScheduler is reliable now, we can introduce (#166) higher latencies for the anti-tracking messages. Also, higher cooldowns.

The idea is mostly to smooth out the traffic that we receive on the server.
Anti-tracking notoriously spikes during 0:00 UTC, and these changes
are intended to decrease the maximum amplitude.